feat: create SingleSelectList Class File
