+++ src/app/src/main/java/jp/kaleidot725/emomemo/ui/common/SingleSelectList.kt
@@ -1,0 +1,22 @@
+package jp.kaleidot725.emomemo.ui.common
+
+class SingleSelectList<T> {
+    private val set: MutableSet<T> = mutableSetOf()
+
+    fun add(item: T) {
+        set.clear()
+        set.add(item)
+    }
+
+    fun clear() {
+        set.clear()
+    }
+
+    fun get(): T {
+        return set.first()
+    }
+
+    fun getList(): List<T> {
+        return set.toList()
+    }
+}
